+++ src/main/java/seedu/duke/command/ViewGradReqCommand.java
@@ -1,6 +1,6 @@
 package seedu.duke.command;
 
-import seedu.duke.list.Grad;
+import seedu.duke.StudyPlan.Grad;
 import seedu.duke.StudyPlan.StudyPlan;
 import seedu.duke.ui.Ui;
 
@@ -13,4 +13,3 @@ public class ViewGradReqCommand extends Command {
         ui.printMessage(requirements);
     }
 }
-
