--- src/main/java/seedu/duke/list/List.java
@@ -1,11 +1,0 @@
-package seedu.duke.list;
-
-public abstract class List {
-
-    protected String code;
-
-    public List(String code) {
-        this.code = code;
-    }
-
-}
