--- src/main/java/seedu/duke/list/Study.java
@@ -1,9 +1,0 @@
-package seedu.duke.list;
-
-public class Study extends List{
-    protected int sem;
-    public Study(String code, int sem) {
-        super(code);
-        this.sem = sem;
-    }
-}
