+++ src/main/java/seedu/duke/ui/Ui.java
@@ -46,34 +46,4 @@ public class Ui {
     public void displaySamplePlan(StudyPlan samplePlan) {
         displayStudyPlan(samplePlan, "CEG Sample Study Plan");
     }
-    //
-    //public void printSamplePlan(){
-    //    System.out.print(line);
-    //    System.out.println("CEG Sample Study Plan");
-    //    System.out.print(line);
-    //    for (Sample sampleItem : sampleItems){
-    //        System.out.println(sampleItem.sem + " " + sampleItem.code);
-    //    }
-    //    System.out.print(line);
-    //}
-    //
-    //public void printGraduationRequirements(){
-    //    System.out.print(line);
-    //    System.out.println("CEG Graduation Requirements");
-    //    System.out.print(line);
-    //    for (Grad gradItem : gradItems){
-    //        System.out.println(gradItem.code);
-    //    }
-    //    System.out.print(line);
-    //}
-    //
-    //public void printStudyPlan{
-    //    System.out.print(line);
-    //    System.out.println("CEG Sample Study Plan");
-    //    System.out.print(line);
-    //    for (Study sampleItem : sampleItems){
-    //        System.out.println(sampleItem.sem + " " + sampleItem.code);
-    //    }
-    //    System.out.print(line);
-    //}
 }
